{{ cookiecutter.project_slug }}/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "{{ cookiecutter.project_slug }}"
version = "0.1.0"
description = ""
authors = ["{{ cookiecutter.copyright_name }}"]
packages = [
    {include = "{{ cookiecutter.package_name }}"},
]
homepage = "https://github.com/{{ cookiecutter.github_user }}/{{ cookiecutter.project_slug }}"
repository = "https://github.com/{{ cookiecutter.github_user }}/{{ cookiecutter.project_slug }}"
readme = "README.md"

[tool.poetry.dependencies]
python = "^{{ cookiecutter.python_version }}"

[tool.poetry.group.dev.dependencies]
autoflake = "^2.2.1"
black = "^23.7.0"
changelog-cmd = "^0.2.0"
invoke = "^2.2.0"
ipdb = "^0.13.13"
ipython = "^8.12"
isort = "^5.12.0"
mypy = "^1.5.1"
pyflakes = "^3.1.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
requests = "^2.31.0"
termcolor = "^2.3.0"
toml = "^0.10"
types-requests = "^2.31.0.2"
types-toml = "^0.10"

[tool.autoflake]
recursive = true
remove-all-unused-imports = true
remove-unused-variables = true
remove-duplicate-keys = true
ignore-init-module-imports = true

[tool.isort]
atomic = true
multi_line_output = 3
profile = "black"
line_length = 120

[tool.black]
line-length = 120
target-version = ['py{{ cookiecutter.python_version.split(".") | join("") }}']
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.mypy]
check_untyped_defs = true
color_output = true
follow_imports = "normal"
show_column_numbers = true
show_error_codes = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = [
    "setuptools.*",
    "invoke.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = "tests"
python_functions = "should_*"
addopts = "-v"
# Structure: https://docs.pytest.org/en/stable/warnings.html
# Values:    https://docs.python.org/3/library/warnings.html#describing-warning-filters
#            action:message:category:module:line
filterwarnings = []
